refactor: share AOT plugin helpers
