+++ micronaut-maven-aot-plugin/pom.xml
@@ -25,7 +25,6 @@
         <dependency>
             <groupId>org.apache.maven</groupId>
             <artifactId>maven-plugin-api</artifactId>
-            <version>${maven.version}</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
@@ -40,7 +39,6 @@
         <dependency>
             <groupId>org.eclipse.sisu</groupId>
             <artifactId>org.eclipse.sisu.inject</artifactId>
-            <version>${sisu.version}</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
